Criação do producer Pessoa, alteração de versão do java e spring boot.

+++ src/main/kotlin/br/com/eventpoc/kafkaschemaregistrypoc/entity/Pessoa.kt
@@ -5,6 +5,6 @@ import java.util.Date
 data class Pessoa(
     val id: String,
     val nome: String,
-    val cpf: Long
+    val cpf: Int
 ) {
 }
